+++ src/components/ui/button/index.ts
@@ -11,7 +11,7 @@ export const buttonVariants = cva(
         default: 'border border-transparent bg-primary text-primary-foreground shadow-xs hover:bg-primary/80',
         primary: 'border border-transparent bg-primary text-primary-foreground shadow-xs hover:bg-primary/80',
         destructive: 'border border-transparent bg-destructive text-white shadow-xs hover:bg-destructive/90 focus-visible:ring-destructive/20 dark:focus-visible:ring-destructive/40 dark:bg-destructive/60',
-        outline: 'border border-border/90 bg-background/20 shadow-xs hover:border-accent-foreground/50 hover:bg-accent hover:text-accent-foreground dark:bg-input/30 dark:border-input dark:hover:bg-input/50',
+        outline: 'text-foreground border border-border/90 bg-background/20 shadow-xs hover:border-accent-foreground/50 hover:bg-accent hover:text-accent-foreground dark:bg-input/30 dark:border-input dark:hover:bg-input/50',
         secondary: 'bg-secondary text-secondary-foreground shadow-xs hover:bg-secondary/80',
         ghost: 'hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent',
         link: 'text-primary underline-offset-4 hover:underline',